renamed and moved keyvault folder to credentials folder in BuildinBlocks
